+++ tests/core.ts
@@ -279,6 +279,7 @@ export const urlToMockRegexp =
 export const testFilePaths = {
     pptx: 'tests/resources/example_presentation_template.pptx',
     docx: 'tests/resources/example_document_template.docx',
+    no_media: 'tests/resources/example_document_template_no_media.docx',
     zip: 'tests/resources/example_zip_template.zip',
 };