+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.20.0] - 2025-10-08
 ### Added
 * Official support for Node.js versions 20, 22 and 24
 ### Fixed
@@ -325,7 +327,8 @@ client library took over this package name. Thanks to
 ownership.
 
 <!-- (Unreleased should never be deleted) -->
-[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.19.1...HEAD
+[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.20.0...HEAD
+[1.20.0]: https://github.com/DeepLcom/deepl-node/compare/v1.19.1...v1.20.0
 [1.19.1]: https://github.com/DeepLcom/deepl-node/compare/v1.19.0...v1.19.1
 [1.19.0]: https://github.com/DeepLcom/deepl-node/compare/v1.18.0...v1.19.0
 [1.18.0]: https://github.com/DeepLcom/deepl-node/compare/v1.17.3...v1.18.0